chore: refactored new itests a bit
